browser_agent/browser_use.py
from typing import Any, Dict, List, Optional
from langchain.schema.language_model import BaseLanguageModel
from langchain_core.callbacks.manager import CallbackManagerForToolRun
from langchain_core.tools import BaseTool
import asyncio
import os

class BrowserTool(BaseTool):
    """Tool for browsing the web and performing actions."""
    
    name = "browser"
    description = "A tool that can browse the web and perform actions in a browser."
    
    def _run(self, task: str, run_manager: Optional[CallbackManagerForToolRun] = None) -> str:
        """Use the tool synchronously."""
        raise NotImplementedError("BrowserTool does not support synchronous execution")
    
    async def _arun(self, task: str, run_manager: Optional[CallbackManagerForToolRun] = None) -> str:
        """Use the tool asynchronously."""
        # This is a simplified implementation
        # In a real implementation, this would use a browser automation library like Playwright or Selenium
        print(f"Browsing the web to: {task}")
        # Simulate web browsing with a delay
        await asyncio.sleep(2)
        
        # This is where you would implement the actual browser automation
        # For this example, we'll return mock data
        if "price" in task.lower() and "gpt-4o" in task.lower() and "deepseek" in task.lower():
            return """
            Research results:
            
            GPT-4o pricing:
            - Input: $10 per 1M tokens
            - Output: $30 per 1M tokens
            
            DeepSeek-V3 pricing:
            - DeepSeek-V3 Coder: $0.50 per 1M tokens input, $1.50 per 1M tokens output
            - DeepSeek-V3 Chat: $0.30 per 1M tokens input, $0.90 per 1M tokens output
            
            Comparison:
            GPT-4o is significantly more expensive than DeepSeek-V3 models.
            """
        
        return "Completed web browsing task, but couldn't find specific information."


class Agent:
    """Agent that can use a browser to complete tasks."""
    
    def __init__(self, task: str, llm: BaseLanguageModel):
        """Initialize the agent with a task and language model."""
        self.task = task
        self.llm = llm
        self.browser_tool = BrowserTool()
    
    async def run(self) -> str:
        """Run the agent on the given task."""
        # In a more sophisticated implementation, this would use LangChain's agent framework
        # to decide when and how to use the browser tool
        
        print(f"Agent received task: {self.task}")
        
        # For this simplified example, we'll just use the browser tool directly
        result = await self.browser_tool._arun(self.task)
        
        # In a real implementation, the LLM would process the browser results
        # and potentially take additional actions
        
        return result
