browser_agent/requirements.txt
langchain-openai>=0.0.5
python-dotenv>=1.0.0
browser-use>=0.1.0
langchain>=0.0.335
